Initial reorg.
