+++ balancer-js/src/modules/swaps/joinExit/actions/index.ts
@@ -1,0 +1,5 @@
+export * from './exit';
+export * from './join';
+export * from './swap';
+export { orderActions, getNumberOfOutputActions } from './helpers';
+export * from './types';
